+++ tests/core/scalar/test_parameter.py
@@ -5,8 +5,8 @@ import pytest
 
 from wetterdienst.metadata.parameter import Parameter
 from wetterdienst.metadata.unit import OriginUnit, SIUnit
-from wetterdienst.provider.dwd.forecast.metadata.parameter import DwdMosmixParameter
-from wetterdienst.provider.dwd.forecast.metadata.unit import DwdMosmixUnit
+from wetterdienst.provider.dwd.mosmix.metadata.parameter import DwdMosmixParameter
+from wetterdienst.provider.dwd.mosmix.metadata.unit import DwdMosmixUnit
 from wetterdienst.provider.dwd.observation.metadata.parameter import (
     DwdObservationDatasetTree,
     DwdObservationParameter,