+++ tests/ui/test_cli.py
@@ -10,7 +10,7 @@ from click.testing import CliRunner
 
 from wetterdienst.ui.cli import cli
 
-# Individual settings for observation and forecast
+# Individual settings for observation and mosmix
 SETTINGS_STATIONS = (
     (
         "dwd",
@@ -21,7 +21,7 @@ SETTINGS_STATIONS = (
     ),
     (
         "dwd",
-        "forecast",
+        "mosmix",
         "--resolution=large --parameter=large",
         "10488",
         "DRESDEN",
@@ -38,7 +38,7 @@ SETTINGS_VALUES = (
     ),
     (
         "dwd",
-        "forecast",
+        "mosmix",
         f"--parameter=small --resolution=large "
         f"--date={datetime.strftime(datetime.today() + timedelta(days=2), '%Y-%m-%d')}",
         "10488",
@@ -64,7 +64,7 @@ def test_cli_about_parameters():
     """Test cli coverage of dwd parameters"""
     runner = CliRunner()
 
-    result = runner.invoke(cli, "about coverage --provider=dwd --kind=observation")
+    result = runner.invoke(cli, "about coverage --provider=dwd --network=observation")
 
     assert "precipitation" in result.output
     assert "temperature_air" in result.output
@@ -75,7 +75,7 @@ def test_cli_about_resolutions():
     """Test cli coverage of dwd resolutions"""
     runner = CliRunner()
 
-    result = runner.invoke(cli, "about coverage --provider=dwd --kind=observation")
+    result = runner.invoke(cli, "about coverage --provider=dwd --network=observation")
 
     assert "minute_1" in result.output
     assert "hourly" in result.output
@@ -85,83 +85,83 @@ def test_cli_about_resolutions():
 def test_cli_about_coverage(capsys):
     runner = CliRunner()
 
-    result = runner.invoke(cli, "about coverage --provider=dwd --kind=observation")
+    result = runner.invoke(cli, "about coverage --provider=dwd --network=observation")
 
     assert "minute_1" in result.output
     assert "precipitation" in result.output
 
 
-def invoke_wetterdienst_stations_empty(provider, kind, setting, fmt="json"):
+def invoke_wetterdienst_stations_empty(provider, network, setting, fmt="json"):
     runner = CliRunner()
 
     return runner.invoke(
         cli,
-        f"stations --provider={provider} --kind={kind} " f"{setting} --station=123456 --format={fmt}",
+        f"stations --provider={provider} --network={network} " f"{setting} --station=123456 --format={fmt}",
     )
 
 
-def invoke_wetterdienst_stations_static(provider, kind, setting, station, fmt="json"):
+def invoke_wetterdienst_stations_static(provider, network, setting, station, fmt="json"):
     runner = CliRunner()
 
     return runner.invoke(
         cli,
-        f"stations --provider={provider} --kind={kind} " f"{setting} --station={station} --format={fmt}",
+        f"stations --provider={provider} --network={network} " f"{setting} --station={station} --format={fmt}",
     )
 
 
-def invoke_wetterdienst_stations_export(provider, kind, setting, station, target):
+def invoke_wetterdienst_stations_export(provider, network, setting, station, target):
     runner = CliRunner()
 
     return runner.invoke(
         cli,
-        f"stations --provider={provider} --kind={kind} " f"{setting} --station={station} --target={target}",
+        f"stations --provider={provider} --network={network} " f"{setting} --station={station} --target={target}",
     )
 
 
-def invoke_wetterdienst_stations_geo(provider, kind, setting, fmt="json"):
+def invoke_wetterdienst_stations_geo(provider, network, setting, fmt="json"):
     runner = CliRunner()
 
     return runner.invoke(
         cli,
-        f"stations --provider={provider} --kind={kind} "
+        f"stations --provider={provider} --network={network} "
         f"{setting} --coordinates=51.1280,13.7543 --rank=5 "
         f"--format={fmt}",
     )
 
 
-def invoke_wetterdienst_values_static(provider, kind, setting, station, fmt="json"):
+def invoke_wetterdienst_values_static(provider, network, setting, station, fmt="json"):
     runner = CliRunner()
 
     return runner.invoke(
         cli,
-        f"values --provider={provider} --kind={kind} " f"{setting} --station={station} --format={fmt}",
+        f"values --provider={provider} --network={network} " f"{setting} --station={station} --format={fmt}",
     )
 
 
-def invoke_wetterdienst_values_export(provider, kind, setting, station, target):
+def invoke_wetterdienst_values_export(provider, network, setting, station, target):
     runner = CliRunner()
 
     return runner.invoke(
         cli,
-        f"values --provider={provider} --kind={kind} " f"{setting} --station={station} --target={target}",
+        f"values --provider={provider} --network={network} " f"{setting} --station={station} --target={target}",
     )
 
 
-def invoke_wetterdienst_values_static_tidy(provider, kind, setting, station, fmt="json"):
+def invoke_wetterdienst_values_static_tidy(provider, network, setting, station, fmt="json"):
     runner = CliRunner()
 
     return runner.invoke(
         cli,
-        f"values --provider={provider} --kind={kind} " f"{setting} --station={station} --format={fmt} --tidy",
+        f"values --provider={provider} --network={network} " f"{setting} --station={station} --format={fmt} --tidy",
     )
 
 
-def invoke_wetterdienst_values_geo(provider, kind, setting, fmt="json"):
+def invoke_wetterdienst_values_geo(provider, network, setting, fmt="json"):
     runner = CliRunner()
 
     return runner.invoke(
         cli,
-        f"values --provider={provider} --kind={kind} {setting} "
+        f"values --provider={provider} --network={network} {setting} "
         f"--coordinates=51.1280,13.7543 --rank=5 --format={fmt}",
     )
 
@@ -169,17 +169,19 @@ def invoke_wetterdienst_values_geo(provider, kind, setting, fmt="json"):
 def test_no_provider():
     runner = CliRunner()
 
-    result = runner.invoke(cli, "stations --provider=abc --kind=abc")
+    result = runner.invoke(cli, "stations --provider=abc --network=abc")
 
     assert "Error: Invalid value for '--provider': 'abc' is not one of 'DWD', 'ECCC', 'NOAA'" in result.output
 
 
-def test_no_kind():
+def test_no_network(caplog):
     runner = CliRunner()
 
-    result = runner.invoke(cli, "stations --provider=dwd --kind=abc")
+    runner.invoke(
+        cli, "stations --provider=dwd --network=abc --parameter=precipitation_height --resolution=daily --all"
+    )
 
-    assert "Invalid value for '--kind': 'abc' is not one of 'OBSERVATION', 'FORECAST'" in result.output
+    assert "No API available for provider DWD and network abc" in caplog.text
 
 
 def test_data_range():
@@ -187,21 +189,21 @@ def test_data_range():
 
     result = runner.invoke(
         cli,
-        "values --provider=eccc --kind=observation --parameter=precipitation_height "
+        "values --provider=eccc --network=observation --parameter=precipitation_height "
         "--resolution=daily --name=toronto",
     )
 
     assert isinstance(result.exception, TypeError)
-    assert "Combination of provider ECCC and kind OBSERVATION requires start and end date" in str(result.exception)
+    assert "Combination of provider ECCC and network OBSERVATION requires start and end date" in str(result.exception)
 
 
 @pytest.mark.parametrize(
-    "provider,kind,setting,station_id,station_name",
+    "provider,network,setting,station_id,station_name",
     SETTINGS_STATIONS,
 )
-def test_cli_stations_json(provider, kind, setting, station_id, station_name):
+def test_cli_stations_json(provider, network, setting, station_id, station_name):
     result = invoke_wetterdienst_stations_static(
-        provider=provider, kind=kind, setting=setting, station=station_id, fmt="json"
+        provider=provider, network=network, setting=setting, station=station_id, fmt="json"
     )
 
     response = json.loads(result.output)
@@ -211,10 +213,10 @@ def test_cli_stations_json(provider, kind, setting, station_id, station_name):
     assert station_name in station_names
 
 
-@pytest.mark.parametrize("provider,kind,setting,station_id,station_name", SETTINGS_STATIONS)
-def test_cli_stations_empty(provider, kind, setting, station_id, station_name, caplog):
+@pytest.mark.parametrize("provider,network,setting,station_id,station_name", SETTINGS_STATIONS)
+def test_cli_stations_empty(provider, network, setting, station_id, station_name, caplog):
 
-    result = invoke_wetterdienst_stations_empty(provider=provider, kind=kind, setting=setting, fmt="json")
+    result = invoke_wetterdienst_stations_empty(provider=provider, network=network, setting=setting, fmt="json")
 
     assert isinstance(result.exception, SystemExit)
     assert "ERROR" in caplog.text
@@ -222,13 +224,13 @@ def test_cli_stations_empty(provider, kind, setting, station_id, station_name, c
 
 
 @pytest.mark.parametrize(
-    "provider,kind,setting,station_id,station_name",
+    "provider,network,setting,station_id,station_name",
     SETTINGS_STATIONS,
 )
-def test_cli_stations_geojson(provider, kind, setting, station_id, station_name):
+def test_cli_stations_geojson(provider, network, setting, station_id, station_name):
 
     result = invoke_wetterdienst_stations_static(
-        provider=provider, kind=kind, setting=setting, station=station_id, fmt="geojson"
+        provider=provider, network=network, setting=setting, station=station_id, fmt="geojson"
     )
 
     response = json.loads(result.output)
@@ -241,30 +243,30 @@ def test_cli_stations_geojson(provider, kind, setting, station_id, station_name)
 
 
 @pytest.mark.parametrize(
-    "provider,kind,setting,station_id,station_name",
+    "provider,network,setting,station_id,station_name",
     SETTINGS_STATIONS,
 )
-def test_cli_stations_csv(provider, kind, setting, station_id, station_name):
+def test_cli_stations_csv(provider, network, setting, station_id, station_name):
 
     result = invoke_wetterdienst_stations_static(
-        provider=provider, kind=kind, setting=setting, station=station_id, fmt="csv"
+        provider=provider, network=network, setting=setting, station=station_id, fmt="csv"
     )
 
     assert station_name in result.output
 
 
 @pytest.mark.parametrize(
-    "provider,kind,setting,station_id,station_name",
+    "provider,network,setting,station_id,station_name",
     SETTINGS_STATIONS,
 )
-def test_cli_stations_excel(provider, kind, setting, station_id, station_name, tmpdir_factory):
+def test_cli_stations_excel(provider, network, setting, station_id, station_name, tmpdir_factory):
 
     # filename = tmpdir_factory.mktemp("data").join("stations.xlsx")  # Noqa:E800
     filename = "stations.xlsx"
 
     _ = invoke_wetterdienst_stations_export(
         provider=provider,
-        kind=kind,
+        network=network,
         setting=setting,
         station=station_id,
         target=f"file://{filename}",
@@ -284,10 +286,10 @@ def test_cli_stations_excel(provider, kind, setting, station_id, station_name, t
     ),
 )
 def test_cli_values_json(setting, expected_columns, capsys, caplog):
-    provider, kind, setting, station_id, station_name = setting
+    provider, network, setting, station_id, station_name = setting
 
     result = invoke_wetterdienst_values_static(
-        provider=provider, kind=kind, setting=setting, station=station_id, fmt="json"
+        provider=provider, network=network, setting=setting, station=station_id, fmt="json"
     )
 
     response = json.loads(result.stdout)
@@ -303,11 +305,11 @@ def test_cli_values_json(setting, expected_columns, capsys, caplog):
     assert set(first.keys()).issuperset(expected_columns)
 
 
-@pytest.mark.parametrize("provider,kind,setting,station_id,station_name", SETTINGS_VALUES)
-def test_cli_values_json_tidy(provider, kind, setting, station_id, station_name):
+@pytest.mark.parametrize("provider,network,setting,station_id,station_name", SETTINGS_VALUES)
+def test_cli_values_json_tidy(provider, network, setting, station_id, station_name):
 
     result = invoke_wetterdienst_values_static_tidy(
-        provider=provider, kind=kind, setting=setting, station=station_id, fmt="json"
+        provider=provider, network=network, setting=setting, station=station_id, fmt="json"
     )
 
     response = json.loads(result.output)
@@ -330,41 +332,41 @@ def test_cli_values_json_tidy(provider, kind, setting, station_id, station_name)
 
 
 @pytest.mark.parametrize(
-    "provider,kind,setting,station_id,station_name",
+    "provider,network,setting,station_id,station_name",
     SETTINGS_STATIONS,
 )
-def test_cli_values_geojson(provider, kind, setting, station_id, station_name, capsys):
+def test_cli_values_geojson(provider, network, setting, station_id, station_name, capsys):
     result = invoke_wetterdienst_values_static(
-        provider=provider, kind=kind, setting=setting, station=station_id, fmt="geojson"
+        provider=provider, network=network, setting=setting, station=station_id, fmt="geojson"
     )
 
     assert "Error: Invalid value for '--format': 'geojson' is not one of 'json', 'csv'" in result.output
 
 
 @pytest.mark.parametrize(
-    "provider,kind,setting,station_id,station_name",
+    "provider,network,setting,station_id,station_name",
     SETTINGS_VALUES,
 )
-def test_cli_values_csv(provider, kind, setting, station_id, station_name):
+def test_cli_values_csv(provider, network, setting, station_id, station_name):
 
     result = invoke_wetterdienst_values_static(
-        provider=provider, kind=kind, setting=setting, station=station_id, fmt="csv"
+        provider=provider, network=network, setting=setting, station=station_id, fmt="csv"
     )
     assert station_id in result.output
 
 
 @pytest.mark.parametrize(
-    "provider,kind,setting,station_id,station_name",
+    "provider,network,setting,station_id,station_name",
     SETTINGS_VALUES,
 )
-def test_cli_values_excel(provider, kind, setting, station_id, station_name, tmpdir_factory):
+def test_cli_values_excel(provider, network, setting, station_id, station_name, tmpdir_factory):
 
     # filename = tmpdir_factory.mktemp("data").join("values.xlsx") # Noqa:E800
     filename = "values.xlsx"
 
     _ = invoke_wetterdienst_values_export(
         provider=provider,
-        kind=kind,
+        network=network,
         setting=setting,
         station=station_id,
         target=f"file://{filename}",
@@ -377,24 +379,24 @@ def test_cli_values_excel(provider, kind, setting, station_id, station_name, tmp
 
 
 @pytest.mark.parametrize(
-    "provider,kind,setting,station_id,station_name",
+    "provider,network,setting,station_id,station_name",
     SETTINGS_VALUES,
 )
-def test_cli_values_format_unknown(provider, kind, setting, station_id, station_name):
+def test_cli_values_format_unknown(provider, network, setting, station_id, station_name):
     result = invoke_wetterdienst_values_static(
-        provider=provider, kind=kind, setting=setting, station=station_id, fmt="foobar"
+        provider=provider, network=network, setting=setting, station=station_id, fmt="foobar"
     )
 
     assert "Error: Invalid value for '--format': 'foobar' is not one of 'json', 'csv'" in result.output
 
 
 @pytest.mark.parametrize(
-    "provider,kind,setting,station_id,station_name",
+    "provider,network,setting,station_id,station_name",
     SETTINGS_STATIONS,
 )
-def test_cli_stations_geospatial(provider, kind, setting, station_id, station_name):
+def test_cli_stations_geospatial(provider, network, setting, station_id, station_name):
 
-    result = invoke_wetterdienst_stations_geo(provider=provider, kind=kind, setting=setting, fmt="json")
+    result = invoke_wetterdienst_stations_geo(provider=provider, network=network, setting=setting, fmt="json")
 
     response = json.loads(result.output)
 
@@ -404,12 +406,12 @@ def test_cli_stations_geospatial(provider, kind, setting, station_id, station_na
 
 
 @pytest.mark.parametrize(
-    "provider,kind,setting,station_id,station_name",
+    "provider,network,setting,station_id,station_name",
     SETTINGS_VALUES,
 )
-def test_cli_values_geospatial(provider, kind, setting, station_id, station_name):
+def test_cli_values_geospatial(provider, network, setting, station_id, station_name):
 
-    result = invoke_wetterdienst_values_geo(provider=provider, kind=kind, setting=setting, fmt="json")
+    result = invoke_wetterdienst_values_geo(provider=provider, network=network, setting=setting, fmt="json")
 
     response = json.loads(result.output)